requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
pydantic>=2.8.0
python-dotenv==1.0.1
psycopg[binary]>=3.2.1
